requirements.txt
scikit-learn==0.22.1
google-cloud-storage==1.25.0
jupyterlab==1.2.6
matplotlib==3.1.3
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
scipy==1.4.1
tqdm==4.42.1
absl-py
pytest
